+++ sdk-java/src/main/java/ly/count/sdk/java/internal/DeviceIdType.java
@@ -4,5 +4,24 @@ package ly.count.sdk.java.internal;
  * Device ID Type
  */
 public enum DeviceIdType {
-    DEVELOPER_SUPPLIED, SDK_GENERATED
+    DEVELOPER_SUPPLIED, SDK_GENERATED;
+
+    public static DeviceIdType fromInt(int deviceIdType, Log L) {
+        switch (deviceIdType) {
+            case 0:
+                return SDK_GENERATED;
+            case 10:
+                return DEVELOPER_SUPPLIED;
+            default:
+                L.e("DeviceIdType, provided int value is not recognized, defaulting to SDK_GENERATED, [" + deviceIdType + "]");
+                return SDK_GENERATED;
+        }
+    }
+
+    public static int toInt(String deviceIdType) {
+        if (deviceIdType.equals("DEVELOPER_SUPPLIED")) {
+            return 10;
+        }
+        return 0;
+    }
 }