+++ sdk-java/src/main/java/ly/count/sdk/java/internal/ModuleDeviceIdCore.java
@@ -185,7 +185,7 @@ public class ModuleDeviceIdCore extends ModuleBase {
     }
 
     /**
-     * Just a wrapper around {@link SDKCore#onSignal(InternalConfig, int, Byteable, Byteable)}} for {@link SDKCore.Signal#DID} case
+     * Just a wrapper around {@link SDKCore#onSignal(InternalConfig, int)}} for {@link SDKCore.Signal#DID} case
      *
      * @param config InternalConfig to run in
      * @param id new {@link Config.DID} if any
@@ -193,7 +193,7 @@ public class ModuleDeviceIdCore extends ModuleBase {
      */
     private void sendDIDSignal(InternalConfig config, Config.DID id, Config.DID old) {
         L.d("[ModuleDeviceIdCore] Sending device id signal: [" + id + "], was [" + old + "]");
-        SDKCore.instance.onSignal(config, SDKCore.Signal.DID.getIndex(), id, old);
+        SDKCore.instance.onSignal(config, SDKCore.Signal.DID.getIndex());
     }
 
     /**
@@ -210,7 +210,8 @@ public class ModuleDeviceIdCore extends ModuleBase {
         } else {
             final Config.DID old = config.getDeviceId();
             config.setDeviceId(new Config.DID(Config.DID.REALM_DID, Config.DID.STRATEGY_CUSTOM, id));
-            Storage.push(config, config);
+            config.storageProvider.setDeviceIdType(DeviceIdType.DEVELOPER_SUPPLIED.name());
+            config.storageProvider.setDeviceID(id);
 
             // old session end & new session begin requests are supposed to happen here
             SDKCore.instance.onDeviceId(config, config.getDeviceId(), old);
@@ -227,7 +228,8 @@ public class ModuleDeviceIdCore extends ModuleBase {
     public void logout(final InternalConfig config) {
         final Config.DID old = config.getDeviceId();
         config.removeDeviceId(old);
-        Storage.push(config, config);
+        config.storageProvider.setDeviceID("");
+        config.storageProvider.setDeviceIdType("");
 
         SDKCore.instance.onDeviceId(config, null, old);
 
@@ -258,7 +260,8 @@ public class ModuleDeviceIdCore extends ModuleBase {
         } else {
             final Config.DID old = config.getDeviceId();
             config.setDeviceId(new Config.DID(Config.DID.REALM_DID, Config.DID.STRATEGY_CUSTOM, id));
-            Storage.push(config, config);
+            config.storageProvider.setDeviceIdType(DeviceIdType.DEVELOPER_SUPPLIED.name());
+            config.storageProvider.setDeviceID(id);
 
             if (withMerge) {
                 clearAndDownloadRCValuesAfterIdChange(false);